Nuke `Notifier`
