+++ app/src/commonMain/kotlin/diy/lingerie/frp/DependentCellVertex.kt
@@ -9,13 +9,13 @@ abstract class DependentCellVertex<V>(
 
     override fun onResumed() {
         subscription.change(
-            strength = Notifier.ListenerStrength.Strong,
+            strength = Vertex.ListenerStrength.Strong,
         )
     }
 
     override fun onPaused() {
         subscription.change(
-            strength = Notifier.ListenerStrength.Weak,
+            strength = Vertex.ListenerStrength.Weak,
         )
     }
