+++ app/src/commonMain/kotlin/diy/lingerie/frp/SwitchCellVertex.kt
@@ -53,7 +53,7 @@ class SwitchCellVertex<V>(
             innerSubscription.cancel()
         }
 
-        override fun change(strength: Notifier.ListenerStrength) {
+        override fun change(strength: Vertex.ListenerStrength) {
             outerSubscription.change(strength = strength)
             innerSubscription.change(strength = strength)
         }
